+++ src/Control/Carrier/State/Strict.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE DeriveFunctor, ExplicitForAll, FlexibleInstances, MultiParamTypeClasses, TypeOperators, UndecidableInstances #-}
+{-# LANGUAGE ExplicitForAll, PatternSynonyms, ViewPatterns #-}
 
 {- | A carrier for the 'State' effect. It evaluates its inner state strictly, which is the correct choice for the majority of use cases.
 
@@ -11,19 +11,16 @@ module Control.Carrier.State.Strict
   runState
 , evalState
 , execState
-, StateC(..)
+, StateC
+, pattern StateC
   -- * State effect
 , module Control.Effect.State
 ) where
 
 import Control.Algebra
-import Control.Applicative (Alternative(..))
 import Control.Effect.State
-import Control.Monad (MonadPlus(..))
-import qualified Control.Monad.Fail as Fail
-import Control.Monad.Fix
-import Control.Monad.IO.Class
-import Control.Monad.Trans.Class
+import qualified Control.Monad.Trans.State.Strict as T
+import Data.Tuple (swap)
 
 -- | Run a 'State' effect starting from the passed value.
 --
@@ -38,7 +35,7 @@ import Control.Monad.Trans.Class
 -- @
 --
 -- @since 1.0.0.0
-runState :: s -> StateC s m a -> m (s, a)
+runState :: Functor m => s -> StateC s m a -> m (s, a)
 runState s (StateC runStateC) = runStateC s
 {-# INLINE[3] runState #-}
 
@@ -66,52 +63,10 @@ execState s = fmap fst . runState s
 
 
 -- | @since 1.0.0.0
-newtype StateC s m a = StateC (s -> m (s, a))
-  deriving (Functor)
+type StateC = T.StateT
 
-instance Monad m => Applicative (StateC s m) where
-  pure a = StateC (\ s -> pure (s, a))
-  {-# INLINE pure #-}
-  StateC f <*> StateC a = StateC $ \ s -> do
-    (s', f') <- f s
-    (s'', a') <- a s'
-    pure (s'', f' a')
-  {-# INLINE (<*>) #-}
-  m *> k = m >>= \_ -> k
-  {-# INLINE (*>) #-}
+pattern StateC :: Functor m => (s -> m (s, a)) -> StateC s m a
+pattern StateC run <- T.StateT (fmap (fmap swap) -> run) where
+  StateC run = T.StateT (fmap swap . run)
 
-instance (Alternative m, Monad m) => Alternative (StateC s m) where
-  empty = StateC (const empty)
-  {-# INLINE empty #-}
-  StateC l <|> StateC r = StateC (\ s -> l s <|> r s)
-  {-# INLINE (<|>) #-}
-
-instance Monad m => Monad (StateC s m) where
-  StateC m >>= f = StateC $ \ s -> do
-    (s', a) <- m s
-    runState s' (f a)
-  {-# INLINE (>>=) #-}
-
-instance Fail.MonadFail m => Fail.MonadFail (StateC s m) where
-  fail s = StateC (const (Fail.fail s))
-  {-# INLINE fail #-}
-
-instance MonadFix m => MonadFix (StateC s m) where
-  mfix f = StateC (\ s -> mfix (runState s . f . snd))
-  {-# INLINE mfix #-}
-
-instance MonadIO m => MonadIO (StateC s m) where
-  liftIO io = StateC (\ s -> (,) s <$> liftIO io)
-  {-# INLINE liftIO #-}
-
-instance (Alternative m, Monad m) => MonadPlus (StateC s m)
-
-instance MonadTrans (StateC s) where
-  lift m = StateC (\ s -> (,) s <$> m)
-  {-# INLINE lift #-}
-
-instance (Algebra sig m, CanHandle sig ((,) s)) => Algebra (State s :+: sig) (StateC s m) where
-  alg (L (Get   k)) = StateC (\ s -> runState s (k s))
-  alg (L (Put s k)) = StateC (\ _ -> runState s k)
-  alg (R other)     = StateC (\ s -> alg (handle (s, ()) (uncurry runState) other))
-  {-# INLINE alg #-}
+{-# COMPLETE StateC #-}